Basic voting system now working across images
A basic voting system has been put in place. Once the user casts a vote,
the system records it and displays it every time the user browse the
same image.
There is no post to the server yet, therefore the current system will
have to be modified to display the average vote after the user casts his
vote.